+++ lib/core.js
@@ -3,26 +3,27 @@ var fs = require('fs');
 var os = require('os');
 var path = require('path');
 var _ = require('lodash');
-var tmp = require('tmp');
-var slash = require('slash');
+//var tmp = require('tmp');
+//var slash = require('slash');
 var penthouse = require('penthouse');
 var CleanCSS = require('clean-css');
 var filterCss = require('filter-css');
 var oust = require('oust');
 var imageInliner = require('imageinliner');
-var request = require('request');
+//var request = require('request');
 /* jshint -W079 */
 var Promise = require('bluebird');
 var tempfile = require('tempfile');
-var url = require('url');
-var mime = require('mime-types');
+//var url = require('url');
+//var mime = require('mime-types');
+var file = require('./fileHelper');
 
 // promisify fs and penthouse
 Promise.promisifyAll(fs);
 var penthouseAsync = Promise.promisify(penthouse);
-var tmpfile = Promise.promisify(tmp.file);
+//var tmpfile = Promise.promisify(tmp.file);
 
-tmp.setGracefulCleanup();
+
 
 /**
  * returns a string of combined and deduped css rules.
@@ -39,132 +40,98 @@ function combineCss(cssArray) {
     ).styles;
 }
 
-/**
- * get path from result array
- * @param resultArray
- * @returns {*}
- */
-function resolveTmp(resultArray) {
-    return _.first(resultArray);
-}
+///**
+// * get path from result array
+// * @param resultArray
+// * @returns {*}
+// */
+//function resolveTmp(resultArray) {
+//    return _.first(resultArray);
+//}
+//
+//function temp(opts) {
+//    return function(resp){
+//        var contentType = resp.headers['content-type'];
+//        return tmpfile(_.assign(opts,{postfix:'.'+mime.extension(contentType)}))
+//            .then(resolveTmp)
+//            .then(function (path) {
+//                return fs.writeFileAsync(path, resp.body).then(function () {
+//                    return path;
+//                });
+//            });
+//    };
+//}
+//
+///**
+// * Fixup slashes in file paths for windows
+// */
+//function normalizePath(str) {
+//    return process.platform === 'win32' ? slash(str) : str;
+//}
+//
+///**
+// * Check wether a resource is external or not
+// * @param href
+// * @returns {boolean}
+// */
+//function isExternal(href) {
+//    return /(^\/\/)|(:\/\/)/.test(href);
+//}
+//
+///**
+// * Get external resource
+// * @param url
+// * @returns {bluebird|exports|module.exports}
+// */
+//function requestAsync(uri) {
+//    return new Promise(function (resolve, reject) {
+//        // handle protocol-relative urls
+//        uri = url.resolve('http://te.st',uri);
+//        request(uri, function (err, resp, body) {
+//            if (err) {
+//                return reject(err);
+//            }
+//            if (resp.statusCode !== 200) {
+//                return reject('Wrong status code ' + resp.statusCode + ' for ' + url);
+//            }
+//            resolve(resp);
+//        });
+//    });
+//}
+//
+///**
+// *
+// * @param base
+// * @returns {bluebird|exports|module.exports}
+// */
+//function assertLocalFile(opts, ext) {
+//    return function (filePath) {
+//        if (!isExternal(filePath)) {
+//            return new Promise(function (resolve) {
+//                resolve(filePath);
+//            });
+//        }
+//
+//        return requestAsync(filePath)
+//            .then(temp({
+//                dir: opts.base
+//            }));
+//    };
+//}
+//
+//function resourcePath(opts){
+//    return function(file){
+//        if (isExternal(file)) {
+//            return file;
+//        }
+//        if (!isExternal(opts.src)){
+//            return path.join(opts.base, file);
+//        }
+//        return url.resolve(opts.src, file);
+//    };
+//}
 
-function temp(opts) {
-    return function(resp){
-        var contentType = resp.headers['content-type'];
-        return tmpfile(_.assign(opts,{postfix:'.'+mime.extension(contentType)}))
-            .then(resolveTmp)
-            .then(function (path) {
-                return fs.writeFileAsync(path, resp.body).then(function () {
-                    return path;
-                });
-            });
-    };
-}
 
-/**
- * Fixup slashes in file paths for windows
- */
-function normalizePath(str) {
-    return process.platform === 'win32' ? slash(str) : str;
-}
-
-/**
- * Check wether a resource is external or not
- * @param href
- * @returns {boolean}
- */
-function isExternal(href) {
-    return /(^\/\/)|(:\/\/)/.test(href);
-}
-
-/**
- * Get external resource
- * @param url
- * @returns {bluebird|exports|module.exports}
- */
-function requestAsync(uri) {
-    return new Promise(function (resolve, reject) {
-        // handle protocol-relative urls
-        uri = url.resolve('http://te.st',uri);
-        request(uri, function (err, resp, body) {
-            if (err) {
-                return reject(err);
-            }
-            if (resp.statusCode !== 200) {
-                return reject('Wrong status code ' + resp.statusCode + ' for ' + url);
-            }
-            resolve(resp);
-        });
-    });
-}
-
-/**
- *
- * @param base
- * @returns {bluebird|exports|module.exports}
- */
-function assertLocalFile(opts, ext) {
-    return function (filePath) {
-        if (!isExternal(filePath)) {
-            return new Promise(function (resolve) {
-                resolve(filePath);
-            });
-        }
-
-        return requestAsync(filePath)
-            .then(temp({
-                dir: opts.base
-            }));
-    };
-}
-
-function resourcePath(opts){
-    return function(file){
-        if (isExternal(file)) {
-            return file;
-        }
-        if (!isExternal(opts.src)){
-            return path.join(opts.base, file);
-        }
-        return url.resolve(opts.src, file);
-    };
-}
-
-/**
- * Get content based on options
- * could either be a html string or a local file
- * @param opts
- * @returns {{promise: *, tmpfiles: Array}}
- */
-function getContentPromise(opts) {
-    if (!(opts.src || opts.html) || !opts.base) {
-        return Promise.reject(new Error('A valid source and base path are required.'));
-    }
-
-    // html passed in directly -> create tmp file and set opts.url
-    if (opts.html) {
-        return tmpfile({dir: opts.base, postfix: '.html'})
-            .then(resolveTmp)
-            .then(function (path) {
-                opts.url = path;
-                return fs.writeFileAsync(opts.url, opts.html).then(function () {
-                    return opts.html;
-                });
-            });
-        // use src file provided
-    } else {
-        return assertLocalFile(opts)(opts.src)
-            .then(function(data) {
-                // src can either be absolute or relative to opts.base
-                if (opts.src !== path.resolve(data) && !isExternal(opts.src)) {
-                    opts.url = path.join(opts.base, opts.src);
-                } else {
-                    opts.url = path.relative(process.cwd(), data);
-                }
-                return fs.readFileAsync(opts.url);
-            });
-    }
-}
 
 
 /**
@@ -190,7 +157,7 @@ function generate(opts) {
 
     return Promise.map(opts.dimensions, function (dimensions) {
         // use content to fetch used css files
-        return getContentPromise(opts).then(function (html) {
+        return file.getContentPromise(opts).then(function (html) {
             // consider opts.css and map to array if it's a string
             if (opts.css) {
                 return typeof opts.css === 'string' ? [opts.css] : opts.css;
@@ -199,9 +166,9 @@ function generate(opts) {
             // Oust extracts a list of your stylesheets (ignoring remote stylesheets)
             var stylesheets = oust(html.toString(), 'stylesheets').filter(function (href) {
                 return true;//!/(^\/\/)|(:\/\/)/.test(href) || isExternal(opts.src) && _.startsWith(href, opts.src);
-            }).map(resourcePath(opts));
+            }).map(file.resourcePath(opts));
 
-            return Promise.map(stylesheets, assertLocalFile(opts));
+            return Promise.map(stylesheets, file.assertLocal(opts));
 
         // read files
         }).map(function (fileName) {
@@ -229,7 +196,7 @@ function generate(opts) {
                     var relativeToBase = path.relative(path.resolve(opts.base), path.resolve(path.join(dir, filePath)));
 
                     var pathPrefix = (typeof opts.pathPrefix === "undefined") ? "/" : opts.pathPrefix;
-                    return normalizePath(match.replace(filePath, path.join(pathPrefix, relativeToBase)));
+                    return file.normalizePath(match.replace(filePath, path.join(pathPrefix, relativeToBase)));
                 });
             });
 
@@ -245,14 +212,13 @@ function generate(opts) {
             });
             // let penthouseAsync do the rest
         }).then(function (csspath) {
-            var pro = penthouseAsync({
-                url: normalizePath(opts.url),
+            return penthouseAsync({
+                url: file.normalizePath(opts.url),
                 css: csspath,
                 // What viewports do you care about?
                 width: dimensions.width,   // viewport width
                 height: dimensions.height  // viewport height
             });
-            return pro;
         });
     }).then(function (criticalCSS) {
         criticalCSS = combineCss(criticalCSS);
@@ -271,7 +237,7 @@ function generate(opts) {
 
 
 exports.generate = generate;
-exports.getContentPromise = getContentPromise;
+//exports.getContentPromise = getContentPromise;
